Refactored AR: PlanningCalendar
The `PlanningCalendar` aggregate root was refactored to capture the
sequence of the items in the calendar (by name).

I gained a better understanding of the MessageDispacther. It only
applied previously saved events to the aggregate, not any events that
are generated from *handling* the current command.

As a result, I should follow this pattern in the handling of commands:
```
// 0) Validation that the command can be processed.
// 1) Create the event object
// 2) "Process" the command by actually trying to do the event, thereby
actually changing the state of the AR (note: any unanticipated
exceptions are wrapped in an InvalidOperationException)
// 3) Return the event (along with any ancillary events that might make
sense, such as ones that provide the more complex state of the object
for a subscriber to better update the Read store)
```